@@ -1,11 +1,11 @@
 {
   "name": "@slate-collaborative/backend",
-  "version": "0.0.3",
+  "version": "0.5.0",
   "files": [
     "lib"
   ],
   "main": "lib/index.js",
-  "types": "lib/model.d.ts",
+  "types": "lib/index.d.ts",
   "description": "slate-collaborative bridge",
   "repository": {
     "type": "git",
@@ -18,20 +18,21 @@
   "license": "MIT",
   "scripts": {
     "prepublishOnly": "yarn run build",
-    "type-check": "tsc --noEmit",
     "build": "yarn run build:types && yarn run build:js",
     "build:types": "tsc --emitDeclarationOnly",
     "build:js": "babel src --out-dir lib --extensions \".ts,.tsx\" --source-maps inline",
     "watch": "yarn build:js -w"
   },
   "dependencies": {
+    "@babel/plugin-proposal-optional-chaining": "^7.9.0",
     "@babel/runtime": "^7.6.3",
-    "@slate-collaborative/bridge": "^0.0.3",
-    "automerge": "^0.12.1",
-    "lodash": "^4.17.15",
-    "slate": "^0.47.8",
-    "socket.io": "^2.2.0",
-    "typescript": "^3.6.3"
+    "@slate-collaborative/bridge": "^0.5.0",
+    "@types/lodash": "^4.14.150",
+    "@types/socket.io": "^2.1.4",
+    "automerge": "^0.14.0",
+    "slate": "^0.57.2",
+    "socket.io": "^2.3.0",
+    "typescript": "^3.8.3"
   },
   "devDependencies": {
     "@babel/cli": "^7.6.0",
@@ -40,8 +41,7 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
     "@babel/plugin-transform-runtime": "^7.6.0",
     "@babel/preset-env": "^7.6.0",
-    "@babel/preset-typescript": "^7.6.0",
-    "@types/socket.io": "^2.1.2"
+    "@babel/preset-typescript": "^7.6.0"
   },
   "directories": {
     "lib": "lib"
