@@ -1,18 +1,17 @@
 {
   "name": "@slate-collaborative/example",
-  "version": "0.0.1",
+  "version": "0.5.0",
   "private": true,
   "dependencies": {
     "@emotion/core": "^10.0.17",
     "@emotion/styled": "^10.0.17",
-    "@slate-collaborative/backend": "^0.0.3",
-    "@slate-collaborative/client": "^0.0.3",
+    "@slate-collaborative/backend": "^0.5.0",
+    "@slate-collaborative/client": "^0.5.0",
     "@types/faker": "^4.1.5",
     "@types/jest": "24.0.18",
     "@types/node": "12.7.5",
-    "@types/react": "16.9.2",
-    "@types/react-dom": "16.9.0",
-    "@types/slate-react": "^0.22.5",
+    "@types/randomcolor": "^0.5.4",
+    "@types/react-dom": "^16.9.6",
     "concurrently": "^4.1.2",
     "cross-env": "^6.0.3",
     "express": "^4.17.1",
@@ -23,14 +22,14 @@
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "react-scripts": "3.1.2",
-    "slate": "^0.47.8",
-    "slate-react": "^0.22.8",
-    "typescript": "3.6.3"
+    "slate": "^0.57.2",
+    "slate-react": "^0.57.2",
+    "typescript": "^3.8.3"
   },
   "scripts": {
     "start": "node server.js",
     "start:cra": "react-scripts start",
-    "build": "cross-env NODE_ENV=production && react-scripts build",
+    "build:site": "cross-env NODE_ENV=production && react-scripts build",
     "dev": "concurrently \"yarn start:cra\" \"yarn serve\"",
     "serve": "nodemon --watch ../backend/lib --inspect server.js"
   },
