@@ -1,8 +1,9 @@
 import * as Automerge from 'automerge'
 
-const createByType = type => (type === 'map' ? {} : type === 'list' ? [] : '')
+const createByType = (type: any) =>
+  type === 'map' ? {} : type === 'list' ? [] : ''
 
-const opCreate = ({ obj, type }: Automerge.Diff, [map, ops]) => {
+const opCreate = ({ obj, type }: Automerge.Diff, [map, ops]: any) => {
   map[obj] = createByType(type)
 
   return [map, ops]
