@@ -1,7 +1,8 @@
 import * as Automerge from 'automerge'
-import { toSlatePath, toJS } from '../utils/index'
 
-const setDataOp = ({ path, value }: Automerge.Diff) => map => ({
+import { toSlatePath, toJS } from '../utils'
+
+const setDataOp = ({ path, value }: Automerge.Diff) => (map: any) => ({
   type: 'set_node',
   path: toSlatePath(path),
   properties: {},
@@ -10,53 +11,19 @@ const setDataOp = ({ path, value }: Automerge.Diff) => map => ({
   }
 })
 
-const AnnotationSetOp = ({ key, value }: Automerge.Diff) => (map, doc) => {
-  if (!doc.annotations) {
-    doc.annotations = {}
-  }
-
-  let op
-
-  /**
-   * Looks like set_annotation option is broken, temporary disabled
-   */
-
-  // if (!doc.annotations[key]) {
-  op = {
-    type: 'add_annotation',
-    annotation: map[value]
-  }
-  // } else {
-  //   op = {
-  //     type: 'set_annotation',
-  //     properties: toJS(doc.annotations[key]),
-  //     newProperties: map[value]
-  //   }
-  // }
-
-  return op
-}
-
 const setByType = {
   data: setDataOp
 }
 
-const opSet = (op: Automerge.Diff, [map, ops]) => {
+const opSet = (op: Automerge.Diff, [map, ops]: any) => {
   const { link, value, path, obj, key } = op
   try {
-    const set = setByType[key]
+    const set = setByType[key as any]
 
     if (set && path) {
       ops.push(set(op))
     } else if (map[obj]) {
-      map[obj][key] = link ? map[value] : value
-    }
-
-    /**
-     * Annotation
-     */
-    if (path && path.length === 1 && path[0] === 'annotations') {
-      ops.push(AnnotationSetOp(op))
+      map[obj][key as any] = link ? map[value] : value
     }
 
     return [map, ops]
