@@ -1,68 +1,28 @@
-import { Selection } from 'slate'
+import { Operation } from 'slate'
+
 import merge from 'lodash/merge'
 
-import { toJS } from '../utils'
-import { SyncDoc, CursorKey } from '../model'
-
 export const setCursor = (
-  doc: SyncDoc,
-  key: CursorKey,
-  selection: Selection,
-  type,
-  data
+  id: string,
+  selection: any = {},
+  doc: any,
+  operations: Operation[],
+  cursorData: any = {}
 ) => {
-  if (!doc) return
+  const cursorOps = operations.filter(op => op.type === 'set_selection')
 
-  if (!doc.annotations) {
-    doc.annotations = {}
-  }
+  if (!doc.cursors) doc.cursors = {}
 
-  if (!doc.annotations[key]) {
-    doc.annotations[key] = {
-      key,
-      type,
-      data: {}
-    }
-  }
+  const newCursor = cursorOps[cursorOps.length - 1]?.newProperties || {}
 
-  const annotation = toJS(doc.annotations[key])
-
-  annotation.focus = selection.end.toJSON()
-  annotation.anchor = selection.start.toJSON()
-
-  annotation.data = merge(annotation.data, data, {
-    isBackward: selection.isBackward,
-    targetPath: selection.isBackward
-      ? annotation.anchor.path
-      : annotation.focus.path
-  })
-
-  doc.annotations[key] = annotation
-
-  return doc
-}
-
-export const removeCursor = (doc: SyncDoc, key: CursorKey) => {
-  if (doc.annotations && doc.annotations[key]) {
-    delete doc.annotations[key]
+  if (selection) {
+    doc.cursors[id] = merge(doc.cursors[id] || {}, newCursor, selection, {
+      ...cursorData,
+      isForward: Boolean(newCursor.focus)
+    })
+  } else {
+    delete doc.cursors[id]
   }
 
   return doc
 }
-
-export const cursorOpFilter = (ops, type: string) =>
-  ops.filter(op => {
-    if (op.type === 'set_annotation') {
-      return !(
-        (op.properties && op.properties.type === type) ||
-        (op.newProperties && op.newProperties.type === type)
-      )
-    } else if (
-      op.type === 'add_annotation' ||
-      op.type === 'remove_annotation'
-    ) {
-      return op.annotation.type !== type
-    }
-
-    return true
-  })
