@@ -1,2 +1,6 @@
-export * from './automerge'
-export * from './slate'
+export type CollabActionType = 'operation' | 'document'
+
+export interface CollabAction {
+  type: CollabActionType
+  payload: any
+}
