@@ -3,7 +3,7 @@ import hexGen from './hexGen'
 
 export * from './testUtils'
 
-const toJS = node => {
+const toJS = (node: any) => {
   try {
     return JSON.parse(JSON.stringify(node))
   } catch (e) {
@@ -12,8 +12,12 @@ const toJS = node => {
   }
 }
 
-const cloneNode = node => toSync(toJS(node))
+const cloneNode = (node: any) => toSync(toJS(node))
 
-const toSlatePath = path => (path ? path.filter(d => Number.isInteger(d)) : [])
+const toSlatePath = (path: any) =>
+  path ? path.filter((d: any) => Number.isInteger(d)) : []
 
-export { toSync, toJS, toSlatePath, hexGen, cloneNode }
+const toCollabAction = (type: any, fn: any) => (payload: any) =>
+  fn({ type, payload })
+
+export { toSync, toJS, toSlatePath, hexGen, cloneNode, toCollabAction }
